requirements.txt
djangorestframework==3.14.0
stripe==5.4.0